% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_inheritance_model.R
\name{apply_inheritance_model}
\alias{apply_inheritance_model}
\title{Apply interval regression model on children inheritance
  to microsimulated data}
\usage{
apply_inheritance_model(
  household_table,
  inheritance_model,
  selection_model = NULL,
  bounds_y_model = NULL,
  deduction = 159325L,
  bounds = c(8072, 12109, 15932, 552324, 902838, 1805677),
  tax_rates = c(5, 10, 15, 20, 30, 40, 45)/100,
  ...
)
}
\arguments{
\item{household_table}{Household longitudinal table}

\item{inheritance_model}{An object estimated by the means
of \link[REtage]{ordered_model_threshold}}

\item{deduction}{Amount deducted from taxation}

\item{bounds}{Thresholds for inheritance tax rates brackets}

\item{tax_rates}{Tax rates}

\item{...}{Additional arguments that should be passed}
}
\value{
Initial dataframe with
\describe{
  \item{y_pred}{Inheritance someone should received given `inheritance_model`
    and family structure (his/her right to bequest)}
}
}
\description{
Apply interval regression model on children inheritance
  to microsimulated data
}
