% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_Destinie.R
\name{arrange_Destinie}
\alias{arrange_Destinie}
\title{Arrange Destinie simulation for longitudinal data}
\usage{
arrange_Destinie(simul)
}
\arguments{
\item{simul}{List of Destinie simulations created by \link{import_Destinie}}
}
\value{
List whose elements are described in \emph{details} section
}
\description{
Create intermediary tables from Destinie's simulation.
}
\details{
The following tables are created:
\itemize{
 \item{\code{pre_indiv}: Individual panel information}
 \item{\code{description}: Individual characteristics}
 \item{\code{macro}: Macro environment variables}
 \item{\code{familles}: Family information}
 \item{\code{pensions}: Individual pensions}
 }
Format (either \emph{dplyr} or \emph{data.table}) depends on the argument
 provided to \link{import_Destinie}
}
\examples{
\dontrun{
path_data <- "./data"
simu <- capitulation::import_Destinie(path_data = path_data)
tables <- capitulation::arrange_Destinie(simu)

# If you prefer data.table approach
simu <- capitulation::import_Destinie(path_data = path_data, format = 'data.table')
tables <- capitulation::arrange_Destinie(simu)
}

}
