% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data_inheritance.R
\name{arrange_inheritance}
\alias{arrange_inheritance}
\title{An intermediate function to arrange microsimulation data
 for inheritance prediction
 
This function aims to arrange microsimulated data in order
 to get the same variable names as *Enquete Patrimoine*}
\usage{
arrange_inheritance(
  household_table,
  data_event,
  tables,
  time_0 = c("graduation", "birth"),
  taille_tr_age = 5,
  taille_tr_agfinetu = 2
)
}
\arguments{
\item{household_table}{Longitudinal data}

\item{data_event}{Data storing inheritance events}

\item{tables}{Destinie's tables}
}
\description{
An intermediate function to arrange microsimulation data
 for inheritance prediction
 
This function aims to arrange microsimulated data in order
 to get the same variable names as *Enquete Patrimoine*
}
