% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_children_R.R
\name{assign_children}
\alias{assign_children}
\title{Match referent with children to finalize household composition}
\usage{
assign_children(table_indiv, descript)
}
\arguments{
\item{table_indiv}{Individual level table, output from
\link{assign_referent}}

\item{descript}{Individual auxiliary information used to determine
if a child is still in household}
}
\value{
\code{table_household} with two additional variables:
\itemize{
 \item{\code{list_children}: comma separated list of children ids}
 \item{\code{UC}: number consumption units at household level}
}
}
\description{
A function to match household husband and spouse information
 with the children they have
}
