% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_referent.R
\name{assign_referent}
\alias{assign_referent}
\title{Assign household head to individual data}
\usage{
assign_referent(table_indiv, fam, descript, call.Rcpp = TRUE)
}
\arguments{
\item{table_indiv}{Dataframe storing individual ids. Output of
\link{prepare}}

\item{fam}{Family information from \link{arrange_Destinie}}

\item{descript}{Personal characteristics table from \link{arrange_Destinie}}

\item{call.Rcpp}{Logical value indicating whether we internally want to use 
\code{C++} (\code{TRUE}) or \code{R} (\code{FALSE}) language}
}
\value{
Individual panel data after being mapped with household head
}
\description{
Starting from individual panel data and family information,
 match individual to a household head
}
\details{
The following rules are applied:
\itemize{
\item{\emph{Immigrants}: Arrive without referent}
\item{\emph{Couples}: Male is assumed to be the head}
\item{\emph{Husband/Spouse death}: Individual is his/her own head}
\item{\emph{Divorce}: Individual is his/her own head}
\item{\emph{Potentially under father/mother authority}:
 If education is accomplished or young adult already earns income,
 individual is his/own head.
Father is assumed to be the household head. If father is dead,
 mother is the head. If she is dead, individual is his/her own head}
}
}
\examples{
\dontrun{
system.time({
capitulation::assign_referent(table_indiv = pre_indiv, fam = simu$fam, descript = description,
                              call.Rcpp = TRUE)
})
# utilisateur     système      écoulé 
#11.300       0.400      11.693 

system.time({
  capitulation::assign_referent(table_indiv = pre_indiv, fam = simu$fam, descript = description,
                                call.Rcpp = FALSE)
})
# utilisateur     système      écoulé 
# 12.844       2.872      22.348 

}
}
\seealso{
\link{assign_referent_cpp} for the internally called
C++ function (recommended one) and \link{assign_referent_R}
for the R version
}
