% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_referent_cpp}
\alias{assign_referent_cpp}
\title{Determine household head for individual level data (C++ version)}
\usage{
assign_referent_cpp(
  id,
  annee,
  conjoint,
  sexe,
  matri,
  referent,
  referent2,
  findet,
  age,
  salaire,
  pere,
  ageMaxPere,
  mere,
  ageMaxMere,
  neFrance
)
}
\arguments{
\item{id}{Individual identifier}

\item{annee}{Observation/Simulation year}

\item{conjoint}{Husband/spouse identifier. Single people
will be imputed a 0 value}

\item{sexe}{Sex. Males are coded 1 and females 2.}

\item{matri}{Matrimonious or family status. \emph{1}: leave parents ;
\emph{2}: couple ; \emph{3}: husband/spouse death ; \emph{4}: divorce}

\item{referent, referent2}{Household heads vector that will be imputed.
Recommended to provide a vector of missing values.}

\item{findet}{Age where inidivudal stops education}

\item{age}{Current age of the individual}

\item{salaire}{Individual income}

\item{pere}{Father identifier}

\item{ageMaxPere}{Father death age}

\item{mere}{Mother identifier}

\item{ageMaxMere}{Mother death age}

\item{neFrance}{Origin of the individual. 1 for individual born
in France ; 0 for immigrants}
}
\value{
Dataframe storing individual identifier (\code{id}),
 year (\code{annee}), matrimonious status (\code{matri}),
 husband/spouse identifier (\code{conjoint}) and referent
 identifier (\code{referent})
}
\description{
\code{Rcpp} function to determine household referent from
 individual data. This is a necessary step to transform individual
 level data into household data. Vector should all be numeric formatted
 and the same length (internally this function uses a for loop).
}
\details{
The following rules are applied:
\itemize{
\item{\emph{Immigrants}: Arrive without referent}
\item{\emph{Couples}: Male is assumed to be the head}
\item{\emph{Husband/Spouse death}: Individual is his/her own head}
\item{\emph{Divorce}: Individual is his/her own head}
\item{\emph{Potentially under father/mother authority}:
 If education is accomplished or young adult already earns income,
 individual is his/own head.
Father is assumed to be the household head. If father is dead,
 mother is the head. If she is dead, individual is his/her own head}
}
}
\seealso{
\link{assign_referent} ; \link{assign_referent_R}
}
