% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_referent_R.R
\name{assign_referent_R}
\alias{assign_referent_R}
\title{Determine household head for individual level data (R version)}
\usage{
assign_referent_R(z, descript)
}
\arguments{
\item{z}{Individual data that should be mapped with household head}

\item{descript}{Personal characteristics table from \link{arrange_Destinie}}
}
\value{
Dataframe storing, between others, individual
 id with household head id
}
\description{
\code{R} function to determine household referent from
 individual data. This is a necessary step to transform individual
 level data into household data. To speed up computations,
 parallelization is performed.
}
\details{
The following rules are applied:
\itemize{
\item{\emph{Immigrants}: Arrive without referent}
\item{\emph{Couples}: Male is assumed to be the head}
\item{\emph{Husband/Spouse death}: Individual is his/her own head}
\item{\emph{Divorce}: Individual is his/her own head}
\item{\emph{Potentially under father/mother authority}:
 If education is accomplished or young adult already earns income,
 individual is his/own head.
Father is assumed to be the household head. If father is dead,
 mother is the head. If she is dead, individual is his/her own head}
}
}
\seealso{
\link{assign_referent} ; \link{assign_referent_cpp} for a
C++ version (recommended one)
}
