% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C0_given_K0_R.R
\name{C0_given_K0}
\alias{C0_given_K0}
\title{Compute initial consumption given initial
 wealth in a life-cycle framework (\code{C++} functions)}
\usage{
C0_given_K0(z, r = 0.02, gamma = 0.5, beta = 1, K0 = 0, timeIndex_var = "tau")
}
\arguments{
\item{z}{Household longitudinal data}

\item{r}{Exogeneous interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Utility discount factor}

\item{K0}{Initial wealth (fixed, estimated or calibrated)}

\item{timeIndex_var}{Time index (t=0,...,T-1) variable}
}
\description{
Compute initial consumption given initial
 wealth in a life-cycle framework (\code{C++} functions)
}
