% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C0_given_K0_noRA}
\alias{C0_given_K0_noRA}
\alias{C0_given_K0_RA}
\alias{C0_given_K0_beta}
\title{Compute initial consumption given initial
 wealth in a life-cycle framework (\code{C++} functions)}
\usage{
C0_given_K0_noRA(income, timeIndex, K0, r)

C0_given_K0_RA(income, timeIndex, K0, r, gamma)

C0_given_K0_beta(income, timeIndex, K0, r, gamma, beta)
}
\arguments{
\item{income}{Income vector}

\item{timeIndex}{Time index (t=0,...,T-1) variable}

\item{K0}{Initial wealth (fixed, estimated or calibrated)}

\item{r}{Exogeneous interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Utility discount factor}
}
\value{
Numerical value representing first period consumption
 in a life-cycle model without risk aversion (\code{C0_given_K0_noRA})
 or with risk aversion (\code{C0_given_K0_RA})
 @param timeIndex_var Time vector of t=0,..,T-1 moments
}
\description{
Compute initial consumption given initial
 wealth in a life-cycle framework (\code{C++} functions)
}
