% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibrate_K0}
\alias{calibrate_K0}
\alias{calibrate_K0_ra}
\alias{calibrate_K0_beta}
\title{Given measured wealth in survey, estimate/calibrate K0 (starting wealth).}
\usage{
calibrate_K0(age, income, findetVector, K2009vector, timeIndex, UC, r)

calibrate_K0_ra(
  age,
  income,
  findetVector,
  K2009vector,
  timeIndex,
  UC,
  r = 0.02,
  gamma = 0.5
)

calibrate_K0_beta(
  age,
  income,
  findetVector,
  K2009vector,
  timeIndex,
  UC,
  r = 0.02,
  gamma = 0.5,
  beta = 1
)
}
\arguments{
\item{age}{Age vector}

\item{income}{Income vector}

\item{findetVector}{End of studying year}

\item{K2009vector}{Measured wealth in survey data expressed as a
vector}

\item{timeIndex}{Time index equals to zero except for
observation year where it should be equal to 1.}

\item{UC}{Number of consumption unit in household at time t=0}

\item{r}{Exogeneous interest rate}

\item{gamma}{Risk-aversion coefficient that gives instantaneous utility function curvature}

\item{beta}{Utility discount rate}
}
\value{
List where first element (\code{K0}) is initial wealth
 and second element (\code{C0}) is life-cycle consumption
}
\description{
Calibrate K0 and C0 such that (i) individual leaves no wealth when he/she dies
 (ii) wealth level is consistent with observed level in 2009 (iii) consumption
 is fixed at C0 level. This function is programmed with \code{C++}.
 \code{calibrate_K0} is designed for risk neutral agents, \code{calibrate_K0_ra}
 for risk adverse agents
}
\details{
In pure permanent income approach with no inheritance, consumption path
 is fixed at a level such that agent does not leave any inheritance. The
 relationship between \eqn{K(t)} and \eqn{K(t-1)} is given by \eqn{K_t = (1+r)(K_{t-1}+Y_t-C_0)}
 with \eqn{C_0 = \frac{K_0+\sum_{t=0}^{T-1} \beta^t Y_t}{\sum_{t=0}^{T-1} \beta^t}}. By
 inverse recursion, K(0) can be determined from K(t)
}
\section{Warning}{

To account for the possibility that familial structure evolves, it is possible
to give the number of consumption unit of the households. In that case,
C0 and K0 are values by consumption unit. To get household level values at
starting time $t=0$,
\eqn{C0*UC(0)$ and $K0*UC(0)} should be considered. To get household level consumption,
at time \eqn{t}, \eqn{C0 \times UC(t)} should be computed
}

\seealso{
\link{simulate_wealth_structural}

\link{simulate_wealth_structural}
}
