% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital_income.R
\name{capital_income}
\alias{capital_income}
\alias{capital_income_heterogeneity}
\title{Compute capital income trajectory from wealth simulations}
\usage{
capital_income(
  simulations,
  wealth_var = "wealth",
  income_var = "y_real",
  scale = c("level", "log", "loglog"),
  r = 0.02
)

capital_income_heterogeneity(
  simulations,
  wealth_var = "wealth",
  income_var = "y_real",
  interest_rate_var = "rvector"
)
}
\arguments{
\item{simulations}{Household level wealth simulations}

\item{wealth_var}{Variable name storing household wealth in \code{simulations}}

\item{income_var}{Individual income column name}

\item{r}{Exogeneous interest rate}

\item{interest_rate_var}{Interest rate vector}
}
\value{
\code{simulations} with variables \code{rK} and \code{rKY} (rK/(rK+w))
}
\description{
Compute capital income trajectory from wealth simulations
}
