% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital_simulation.R
\name{capital_simulation}
\alias{capital_simulation}
\title{Wealth simulations based on life-cycle approach}
\usage{
capital_simulation(
  data,
  method = "structural",
  start_year = 2009,
  estimation_year = 2009,
  final_year = 2070,
  id_var = "Id",
  income_var = "salaire",
  age_var = "age",
  ageliq_var = "ageliq",
  matrimonial_var = "matri",
  findet_var = "findet",
  wealthvar_survey = "PATFISOM",
  deflate = TRUE,
  r = 0.03,
  gamma = 0.5,
  beta = 0.98,
  verbose = TRUE,
  weight_var = NULL,
  return_last = FALSE
)
}
\arguments{
\item{data}{If a dataset has already been prepared for microsimulation
use it to save time. Otherwise set \code{data} to \code{NULL} and
\link{prepare_data} will be used to import and reshape Destinie
simulations}

\item{method}{Modelization adopted. For the moment, only
\emph{structural} approach is implemented}

\item{start_year}{Starting year}

\item{estimation_year}{Estimation year in wealth survey}

\item{id_var}{Individual household identifier column name}

\item{income_var}{Income variable in microsimulated data}

\item{age_var}{Individual age column name}

\item{ageliq_var}{Individual liquidation age variable name}

\item{matrimonial_var}{Matrimonial status variable}

\item{findet_var}{End of studying year variable}

\item{wealthvar_survey}{Variable for wealth in survey data}

\item{deflate}{Logical value indicating whether values
should be deflated}

\item{r}{Exogeneous interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Discount factor}

\item{verbose}{Logical value indicating whether we want to print some messages
during computations}

\item{weight_var}{Variable name to weight observations. 
By default, set to \code{NULL} to consider individual data}

\item{return_last}{Logical value indicating whether we should
return \eqn{Kt} sequence up to `T-1` (default) or `T`}
}
\value{
\link[data.table]{data.table} object with
 wealth simulations between \code{start_year} and
 \code{final_year}
}
\description{
This global function (i) shapes Destinie data if needed
 (ii) estimate initial wealth for first generation
 (iii) simulate wealth for first and following generations
}
