% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_model.R
\name{constructor_interest}
\alias{constructor_interest}
\alias{constructor_gamma}
\alias{constructor_beta}
\alias{constructor_interest.default}
\alias{constructor_interest.double}
\alias{constructor_interest.formula}
\alias{constructor_interest.character}
\title{Construct interest rate model in a heterogeneous or homogeneous agents model}
\usage{
constructor_interest(r, data, r.parameters = NULL, ...)

constructor_gamma(gamma, data, gamma.parameters = NULL, ...)

constructor_beta(beta, data, beta.parameters = NULL, ...)

\method{constructor_interest}{default}(r, data, r.parameters = NULL, ...)

\method{constructor_interest}{double}(r, data, r.parameters = NULL, ...)

\method{constructor_interest}{formula}(r, data, r.parameters = NULL, ...)

\method{constructor_interest}{character}(r, data, r.parameters = NULL, ...)
}
\arguments{
\item{r}{A numeric value or a formula applied to data}

\item{data}{Dataframe where r should be created}

\item{r.parameters}{Parameters value that should be used if
r is a formula}

\item{...}{Additional arguments that should be passed}

\item{gamma}{A numeric value or a formula applied to data}

\item{gamma.parameters}{Parameters value that should be used if
gamma is a formula}

\item{beta}{A numeric value or a formula applied to data}

\item{beta.parameters}{Parameters value that should be used if
beta is a formula}
}
\description{
Construct interest rate model in a heterogeneous or homogeneous agents model
}
