% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_event.R
\name{death_event}
\alias{death_event}
\title{Create dataframe storing heir share to family
 wealth when a death event occurs}
\usage{
death_event(df_longitudinal)
}
\arguments{
\item{df_longitudinal}{Longitudinal data taking the form
of the output of \link{household_composition}}
}
\value{
A dataframe with following variables
\describe{
  \item{Id}{Identifier for the individual that dies}
  \item{annee}{Year at which event occurs}
  \item{share}{Share that heir will have}
  \item{heir}{Identifier for heir}
}
}
\description{
Create dataframe storing heir share to family
 wealth when a death event occurs
}
