% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_K0.R
\name{estimate_K0}
\alias{estimate_K0}
\title{Estimate initial wealth (K0) that is consistent with observed
 value in survey in a pure life-cycle model}
\usage{
estimate_K0(
  data,
  r = 0.02,
  beta = 1,
  gamma = 0.5,
  wealthvar_survey = "PATFISOM",
  observation_year = 2009,
  income_var = "salaire_tot",
  weight_var = NULL
)
}
\arguments{
\item{data}{Microsimulated data merged with household data
in the relevent observation year}

\item{r}{Interest rate}

\item{beta}{Discount factor}

\item{gamma}{Risk aversion coefficient}

\item{wealthvar_survey}{Variable for wealth in survey data}

\item{observation_year}{Survey data year}

\item{income_var}{Income variable in microsimulated data}

\item{weight_var}{Variable storing individual weight inside household. This can be 1
(individual is the unique adult representing household), 2 (couples) or
a value related to consumption units}
}
\description{
Estimate initial wealth (K0) that is consistent with observed
 value in survey in a pure life-cycle model
}
