% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_K0_old}
\alias{fit_K0_old}
\title{Determine K0 in a life-cycle model.}
\usage{
fit_K0_old(income, K2009vector, timeIndex, UC, r = 0.02, gamma = 0.5, beta = 1)
}
\arguments{
\item{income}{Income vector}

\item{K2009vector}{Observed wealth in survey data}

\item{timeIndex}{A 0/1 vector where 1 indicates
year where individual is observed on wealth data}

\item{UC}{Number of consumption units}

\item{r}{Interest rate vector. Because we assume an individual
faces the same interest rate every year, only the first value
is used}

\item{gamma}{Risk-aversion coefficient}

\item{beta}{Discount factor}
}
\description{
Determine K0 in a life-cycle model.
}
