% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pre_indiv.R
\name{generate_fake_data}
\alias{generate_fake_data}
\title{Generate fake \code{Destinie} data}
\usage{
generate_fake_data(Nindiv = 4, anaiss = NULL, lifetime = NULL, findet = NULL)
}
\arguments{
\item{Nindiv}{Number of individuals}

\item{anaiss}{Birth years}

\item{lifetime}{Lifetime for individuals in dataset}

\item{findet}{End of studying year vector}
}
\description{
Generate fake \code{Destinie} data
}
