% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pre_indiv.R
\name{generate_pre_indiv}
\alias{generate_pre_indiv}
\title{Generate a fake \code{pre_indiv} table}
\usage{
generate_pre_indiv(
  Nindiv = 2,
  anaiss = c(1980, 1990),
  lifetime = c(50, 60),
  findet = c(25, 22)
)
}
\arguments{
\item{Nindiv}{Number of individuals}

\item{anaiss}{Birth years}

\item{lifetime}{Lifetime for individuals in dataset}

\item{findet}{End of studying year vector}
}
\value{
Dataframe
}
\description{
Generate a fake \code{pre_indiv} table
}
