% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_composition.R
\name{household_composition}
\alias{household_composition}
\title{Associate individuals with the family they belong to}
\usage{
household_composition(table_indiv, simul, descript, call.Rcpp = TRUE)
}
\arguments{
\item{table_indiv}{Dataframe storing individual ids. Output of
\link{prepare}}

\item{simul}{Destinie tables stored as a list}

\item{descript}{Personal characteristics table from \link{arrange_Destinie}}

\item{call.Rcpp}{Logical value indicating whether we internally want to use 
\code{C++} (\code{TRUE}) or \code{R} (\code{FALSE}) language}
}
\description{
This function is designed to associate individual with
\enumerate{
  \item Household head or heads (couples)
  \item Children
}
}
