% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Destinie.R
\name{import_Destinie}
\alias{import_Destinie}
\title{Import Destinie demographic simulations}
\source{
<https://github.com/InseeFr/Destinie-2>
}
\usage{
import_Destinie(
  filenames = c("emp", "salairenet", "ech", "macro", "fam", "liquidations", "retraites"),
  path_data = getwd(),
  format = c("dplyr", "data.table"),
  extension = ".Rda"
)
}
\arguments{
\item{filenames}{Names of the files, excluding extension}

\item{path_data}{Directory where files should be found}

\item{format}{Should we use \code{tidyverse} approach (format = "dplyr") or
\code{data.table} approach (format = "data.table"). Default to "\emph{dplyr}"}
}
\value{
List where each slot stores a \link[tibble]{tibble} or a
 \link[data.table]{data.table} objects with simulations
 from Destinie
}
\description{
Import simulations from Destinie model
}
\details{
The following output from INSEE's Destinie model are required:
\itemize{
 \item{\code{emp}: Individual identifiers} 
 \item{\code{salairenet}: Simulated individual income path}
 \item{\code{ech}: Simulated family structure}
 \item{\code{macro}: Macroeconomic variables (simulated or observed)}
 \item{\code{fam}: Another table regarding family links}
 \item{\code{liquidations}: Simulated retirement decision}
 \item{\code{retraites}: Simulated pensions}
 }
 
 They should be stored as \code{.Rda} files
}
\note{
It is recommended to favor \link{data.table} approach though it is not,
 for compability and continuity purposes, the default approach. This choice
 determines following steps: all functions have been built to adapt computations
 with format required
}
\examples{
\dontrun{
path_data <- "./data"
simu <- capitulation::import_Destinie(path_data = path_data)
tables <- capitulation::arrange_Destinie(simu)

# If you prefer data.table approach
simu <- capitulation::import_Destinie(path_data = path_data, format = 'data.table')
tables <- capitulation::arrange_Destinie(simu)
}
}
\seealso{
\link{arrange_Destinie} to get intermediary tables
}
