% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_group.R
\name{income_group}
\alias{income_group}
\title{Income group an individual belongs to}
\usage{
income_group(
  data,
  quantile = 0.9,
  idvar = "Id",
  groupvar = "top",
  incomevar = "y_indiv",
  yearvar = "annee"
)
}
\arguments{
\item{data}{A \link[data.table]{data.table} object}

\item{quantile}{Quantile used (between 0 and 1) to cut
population in two classes}

\item{idvar}{Id variable name}

\item{groupvar}{Variable prefix for the group identification.
Variable name will be \code{group_var}_(1-\code{quantile})}

\item{incomevar}{Income variable}

\item{yearvar}{Year variable. Used for \code{by} statement}
}
\description{
Determine to which income group an individual belongs to
}
\details{
This
}
