% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_household.R
\name{income_household}
\alias{income_household}
\title{Aggregate income at household level}
\usage{
income_household(
  table_indiv,
  id_var = "Id",
  income_var = "salaire",
  age_var = "age",
  ageliq_var = "ageliq",
  matrimonial_var = "matri"
)
}
\arguments{
\item{table_indiv}{Individual level data where
people are matched to husband/spouse id}

\item{id_var}{Individual household identifier column name}

\item{income_var}{Individual income column name}

\item{age_var}{Individual age column name}

\item{ageliq_var}{Individual liquidation age variable name}

\item{matrimonial_var}{Matrimonial status variable}
}
\value{
Individual level data with income aggregated
 at household level
}
\description{
Match individual with income at household level
}
\details{
By assumption, male is assumed to be the household head.

Several arguments are proposed to ensure a column name discrepancy
 between two inputs will not produce different results
}
