% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inherit_wealth}
\alias{inherit_wealth}
\title{Wealth transmission from year T-1 to T}
\usage{
inherit_wealth(
  id,
  patrimoine,
  year,
  matri_status,
  sexe,
  spouse_id,
  patrimoine_conjoint,
  matri_conjoint,
  year_loop,
  referent_list
)
}
\arguments{
\item{id}{Individual identifier}

\item{patrimoine}{Wealth that has been
imputed in a previous iteration 
or that should be imputed before next
iteration}

\item{year}{Year vector}

\item{matri_status}{Matrimonious or family status. \emph{1}: leave parents ;
\emph{2}: couple ; \emph{3}: husband/spouse death ; \emph{4}: divorce}

\item{sexe}{Sex. Males are coded 1 and females 2.}

\item{spouse_id}{Husband/Spouse identifier. Missing values if no
husband/spouse}

\item{patrimoine_conjoint}{Husband/Spouse wealth}

\item{matri_conjoint}{Husband/Spouse matrimonious or family status. \emph{1}: leave parents ;
\emph{2}: couple ; \emph{3}: husband/spouse death ; \emph{4}: divorce}

\item{year_loop}{T: loop year inside \code{inherit_wealth} is launched}

\item{referent_list}{List of household head in T-1}
}
\value{
List: K[T-1], K[T] after demographic change
}
\description{
Take last imputed year (T-1) and transmit
 wealth to year T given household change
}
\details{
Among the most remarkable rules:
\itemize{
\item{\emph{Divorce}: Household wealth divided between spouses}
\item{\emph{Marriage}: Household wealth merged between spouses}
}
}
