% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_year_change.R
\name{last_change_indiv}
\alias{last_change_indiv}
\title{Determine last family change}
\usage{
last_change_indiv(household_table)
}
\arguments{
\item{household_table}{Longitudinal household table}
}
\value{
Initial table with \code{last_year} variable giving
 year of last change
}
\description{
Determine when was last family change (marriage, divorce, etc.)
 an individual has known. The breaks of family change give us information
 on the years where household wealth should be splitted/merged between individuals
}
\details{
This step is necessary to ease microsimulation computations
 since family change provoke splitting and merging individual wealth
}
