% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{life_cycle_apply}
\alias{life_cycle_apply}
\title{Life cycle model}
\usage{
life_cycle_apply(
  input,
  f,
  id_var = "Id",
  annee_var = "annee",
  income_var = "salaire_tot",
  observed_wealth_var = "wealth2009",
  time_var = "tt",
  weight_var = "UC",
  Hgiven_var = "tempHg",
  Hreceived_var = "tempHr",
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  returnLast = FALSE
)
}
\arguments{
\item{input}{List of data}

\item{f}{Function to apply many times}

\item{id_var}{Identifier variable name}

\item{annee_var}{Time vector name}

\item{income_var}{Income variable name}

\item{observed_wealth_var}{Observed wealth variable name}

\item{time_var}{0-1 variable name with 1 for 2009 value}

\item{weight_var}{Weight variable. See \link{prepare_data}}

\item{Hgiven_var}{Inheritance given variable}

\item{Hreceived_var}{Inheritance received variable}

\item{r}{Interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Discount factor}

\item{returnLast}{Return last element}
}
\value{
A list of dataframe
}
\description{
Life cycle model
}
