% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_cycle_model.R
\name{life_cycle_model}
\alias{life_cycle_model}
\title{Life cycle model}
\usage{
life_cycle_model(
  data,
  r = 0.02,
  beta = 1,
  gamma = 0.5,
  alpha = NULL,
  r.parameters = NULL,
  gamma.parameters = NULL,
  beta.parameters = NULL,
  non_ricardian = FALSE,
  observation_year = 2009,
  use_consumption = FALSE,
  id_var = "Id",
  year_var = "annee",
  wealthvar_survey = "K_observed",
  income_var = "y_indiv",
  age_var = "age",
  output_var = "wealth",
  probability_survival_var = NULL,
  non_ricardian_var = NULL,
  weight_var = NULL,
  get_capital_income = FALSE,
  return_last = FALSE,
  additional_vars = NULL,
  scale_model = c("level", "log", "loglog"),
  ...
)
}
\arguments{
\item{data}{Dataframe where r should be created}

\item{r}{Interest rate vector. Because we assume an individual
faces the same interest rate every year, only the first value
is used}

\item{beta}{Discount factor}

\item{gamma}{Risk-aversion coefficient}

\item{r.parameters}{Parameters value that should be used if
r is a formula}

\item{gamma.parameters}{Parameters value that should be used if
gamma is a formula}

\item{beta.parameters}{Parameters value that should be used if
beta is a formula}

\item{non_ricardian}{Should we introduce some non-Ricardian household}

\item{observation_year}{Survey data year}

\item{id_var}{Variable name for identifier}

\item{year_var}{Variable name for time vector}

\item{wealthvar_survey}{Variable for wealth in survey data}

\item{income_var}{Income variable in microsimulated data}

\item{age_var}{Age variable}

\item{output_var}{Wealth variable name that should be used}

\item{weight_var}{Variable storing individual weight inside household. This can be 1
(individual is the unique adult representing household), 2 (couples) or
a value related to consumption units}

\item{get_capital_income}{Logical value indicating whether
we should deduce capital income from wealth}

\item{return_last}{Logical value indicating whether we want the sequence
of wealth to be \eqn{\{K_0,...,K_{T-1}\}} (\code{FALSE}, default)
or  \eqn{\{K_1,...,K_{T}\}} (\code{TRUE})}

\item{...}{Additional arguments that are currently ignored by \code{life_cycle_model}}
}
\description{
These functions are \code{R}-level function
 to run life-cycle model
}
\details{
\code{life_cycle_model} for the homogeneous interest
 rate model, \code{life_cycle_model_heterogeneity} for the
 individual varying interest rate
}
