% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{life_cycle_model_cpp}
\alias{life_cycle_model_cpp}
\title{Estimate wealth accumulation with a life-cycle model}
\usage{
life_cycle_model_cpp(
  income,
  observed_wealth,
  pi,
  timeIndex,
  inheritanceGiven,
  inheritanceReceived,
  r,
  risk_aversion,
  discount_factor,
  scale_model = "level",
  outcome = "wealth"
)
}
\arguments{
\item{income}{Income vector}

\item{observed_wealth}{Observed wealth in survey data or
stored into long vector (i.e. a vector where each
row presents the same value). Missing values are not
allowed}

\item{pi}{Surival probability vector}

\item{timeIndex}{A 0/1 vector where 1 indicates
year where individual is observed on wealth data}

\item{inheritanceGiven}{Vector storing inheritance that
an individual wants to transmit when dying. Stored into
long vector (i.e. a vector where each
row presents the same value). Missing values are not
allowed}

\item{inheritanceReceived}{Vector storing inheritance that
an individual receives every year. Zero values when no
inheritance is received and positive amount when
the individual receives.}

\item{r}{Interest rate vector. Because we assume an individual
faces the same interest rate every year, only the first value
is used}

\item{gamma}{Risk-aversion coefficient}

\item{beta}{Discount factor}
}
\value{
The sequence of an individual wealth: \eqn{(K_t)_t} vector
}
\description{
Given individualized income trajectory, one
observed wealth point in time and a model prediction for
inheritance,
determine wealth trajectory in a life-cycle model.
This implementation uses the closed form formula
presented in the paper.
}
\details{
These functions are designed to determine
 the full individual wealth dynamics given income
 trajectory. It is assumed that an observed wealth
 vector \code{observed_wealth} is provided where only
 one value is meaningful, the one where \code{timeIndex}
 is equal to one (observation dummy).
 These functions implement Modigliani-Friedman model 
 augmented with inheritance in a context of
 microsimulation where every year income is observed
}
