% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{life_cycle_model_cpp_bis_uncertainty}
\alias{life_cycle_model_cpp_bis_uncertainty}
\title{Same function than life_cycle_model_cpp with the
 additional possibility of having K_{T+1}}
\usage{
life_cycle_model_cpp_bis_uncertainty(
  income,
  observed_wealth,
  pi,
  timeIndex,
  inheritanceGiven,
  inheritanceReceived,
  r,
  risk_aversion,
  discount_factor,
  scale_model = "level",
  return_last = FALSE
)
}
\arguments{
\item{income}{Income vector}

\item{observed_wealth}{Observed wealth in survey data or
stored into long vector (i.e. a vector where each
row presents the same value). Missing values are not
allowed}

\item{pi}{Surival probability vector}

\item{timeIndex}{A 0/1 vector where 1 indicates
year where individual is observed on wealth data}

\item{inheritanceGiven}{Vector storing inheritance that
an individual wants to transmit when dying. Stored into
long vector (i.e. a vector where each
row presents the same value). Missing values are not
allowed}

\item{inheritanceReceived}{Vector storing inheritance that
an individual receives every year. Zero values when no
inheritance is received and positive amount when
the individual receives.}

\item{r}{Interest rate vector. Because we assume an individual
faces the same interest rate every year, only the first value
is used}

\item{return_last}{Logical value indicating whether
we would like to return last value}
}
\description{
Same function than life_cycle_model_cpp with the
 additional possibility of having K_{T+1}
}
