% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{life_cycle_model_cpp_old}
\alias{life_cycle_model_cpp_old}
\alias{life_cycle_model_cpp2_old}
\title{[Discarded] Estimate wealth accumulation with a life-cycle model}
\usage{
life_cycle_model_cpp_old(
  income,
  K2009vector,
  timeIndex,
  UC,
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  returnLast = FALSE
)

life_cycle_model_cpp2_old(
  income,
  K2009vector,
  timeIndex,
  UC,
  inheritanceGiven,
  inheritanceReceived,
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  returnLast = FALSE
)
}
\arguments{
\item{income}{Income vector}

\item{K2009vector}{Observed wealth in survey data or
missing value (in that case K0 will be 0) stored
into long vector}

\item{timeIndex}{A 0/1 vector where 1 indicates
year where individual is observed on wealth data}

\item{UC}{Number of consumption unit in household at time t=0}

\item{r}{Interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Discount factor}

\item{returnLast}{Return last element}

\item{inheritanceGiven}{Vector storing inheritance that
an individual wants to transmit when dying. Stored into
long vector (i.e. a vector where each
row presents the same value). Missing values are not
allowed}

\item{inheritanceReceived}{Vector storing inheritance that
an individual receives every year. Zero values when no
inheritance is received and positive amount when
the individual receives.}
}
\description{
See \link{life_cycle_model} for the current implementation.
Given individualized income trajectory and one
observed wealth point in time,
determine wealth trajectory in a life-cycle model
(see \code{details}). People whom wealth is
not observed at any time are assumed to start
without wealth (\eqn{K_0 = 0}). Functions differ
in the structure of input required or output returned
}
\details{
These functions are designed to determine
 the full individual wealth dynamics given income
 trajectory. It is assumed that an observed wealth
 vector \code{K2009vector} is provided where only
 one value is meaningful, the one where \code{timeIndex}
 is equal to one (observation dummy).
 These functions implementat of
 Modigliani-Friedman model in a context of
 microsimulation where every year income is observed
\describe{
  \item{life_cycle_model_cpp}{Interest rate
  as a parameter. Return as a list:
  \itemize{
    \item \eqn{K_0} consistent with \eqn{K_t} and
     model hypotheses
    \item \eqn{C_0} consistent with \eqn{K_0}
    \item \eqn{K_t} sequence
    }
  }
  \item{life_cycle_model_cpp2}{Same but only returns
   \eqn{K_t} sequence}
  \item{life_cycle_model_heterogeneity_cpp}{Interest rate
  as a vector. Return as a list:
  \itemize{
    \item \eqn{K_0} consistent with \eqn{K_t} and
     model hypotheses
    \item \eqn{C_0} consistent with \eqn{K_0}
    \item \eqn{K_t} sequence
    }
  }
}
}
