% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_cycle_model_parallel.R
\name{life_cycle_model_parallel}
\alias{life_cycle_model_parallel}
\title{Life cycle model parallelized at \code{R} level}
\usage{
life_cycle_model_parallel(
  data,
  r = 0.02,
  beta = 1,
  gamma = 0.5,
  observation_year = 2009,
  id_var = "Id",
  year_var = "annee",
  age_var = "age",
  wealthvar_survey = "PATFISOM",
  income_var = "salaire_tot",
  weight_var = NULL,
  return_last = FALSE,
  get_capital_income = FALSE
)
}
\arguments{
\item{return_last}{Logical value indicating whether we want the sequence
of wealth to be \eqn{\{K_0,...,K_{T-1}\}} (\code{FALSE}, default)
or  \eqn{\{K_1,...,K_{T}\}} (\code{TRUE})}
}
\description{
This function is the \code{R}-level function
 to run life-cycle model
}
