% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\docType{data}
\name{macro}
\alias{macro}
\title{Price index from 1900 to 2070}
\format{
A \link[data.table]{data.table} object with 171 rows and 2 variables:
\describe{
  \item{annee}{Year}
  \item{Prix}{Price index}
}
}
\source{
\url{https://github.com/InseeFr}
}
\usage{
macro
}
\description{
Longitudinal dataset for price index used to
 deflate data
}
\keyword{datasets}
