% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_wealthSurvey.R
\name{match_wealthSurvey}
\alias{match_wealthSurvey}
\title{Match wealth household survey with Destinie data}
\usage{
match_wealthSurvey(
  data,
  path_survey = "./inst/dataINSEE",
  filename_survey = "basecomplete0_62663.sas7bdat",
  level_matching = c("household", "individual"),
  id_wealthSurvey = "IDENTMEN",
  age_wealthSurvey = "AGEPR",
  wealth_wealthSurvey = "PATFISOM",
  debt_wealthSurvey = NULL,
  findet_wealthSurvey = "AGFINETU",
  .colsWealth = c("IDENTIND", "IDENTMEN", "age", "AGEPR", "POND", "MTDETTES",
    "PATRI_NET", "PATRI_BRUT", "PATFI", "PATFIMTC_DECL", "PATFISOM", "PATIMM",
    "PATPROFENT", "PATPROFHENT", "PATRIC_DECL", "AGFINETU"),
  estimation_year = 2009L,
  time_0 = c("graduation", "birth"),
  drawK = FALSE
)
}
\arguments{
\item{data}{Destinie data}

\item{path_survey}{Directory where survey data are stored}

\item{filename_survey}{Survey data filename}

\item{level_matching}{Should we match at individual or household level. For the moment,
only household level is implemented}

\item{id_wealthSurvey}{Identifier in household survey}

\item{age_wealthSurvey}{Age in household survey}

\item{wealth_wealthSurvey}{Wealth in household survey}

\item{findet_wealthSurvey}{Studying end age in household survey}

\item{.colsWealth}{Columns of interest in survey data}

\item{estimation_year}{Estimation year in wealth survey}

\item{drawK}{Logical value indicating whether we would like to draw observed value
(wealth, age) for people that have no wealth observed in \code{estimation_year}}
}
\value{
\code{data} augmented with 'Enquete Patrimoine' information for 2009 year
}
\description{
Match simulated household in Destinie with their counterpart in
 observed data (household survey 'Enquete Patrimoine').
}
