% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_K0}
\alias{param_K0}
\title{Intermediate function to compute some parameters}
\usage{
param_K0(r, gamma, beta, income, UC)
}
\arguments{
\item{r}{Interest rate vector. Because we assume an individual
faces the same interest rate every year, only the first value
is used}

\item{gamma}{Risk-aversion coefficient}

\item{beta}{Discount factor}

\item{income}{Income vector}

\item{UC}{Number of consumption units}
}
\description{
Intermediate function to compute some parameters
}
