% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gini.R
\name{plot_gini}
\alias{plot_gini}
\title{Plot Gini index evolution on microsimulated series}
\usage{
plot_gini(
  simulations,
  year_var = "annee",
  vars = c("y_indiv", "wealth", "Y"),
  labels = c("Labor income", "Financial wealth", "Total income"),
  start_year = 2009,
  final_year = 2040,
  export_plot = FALSE,
  export_path = getwd(),
  export_filename = "04_gini",
  langage = c("French", "English"),
  scale_viridis = FALSE,
  negative_values = c("truncate", "keep", "remove"),
  ...
)
}
\arguments{
\item{simulations}{Longitudinal household level dataframe
storing wealth simulations}

\item{year_var}{Year variable}

\item{vars}{Variables that should be represented}

\item{labels}{Variable labels for legend}

\item{start_year}{First year of simulation}

\item{final_year}{Last year of simulation}

\item{export_plot}{Logical value indicating whether we want
to export the plot (\code{TRUE}, default) or not}

\item{export_path}{Path to export data}

\item{export_filename}{Filename if data are exported. PDF extension
will be added if missing}

\item{langage}{Should we label axes in English or French ?}

\item{scale_viridis}{Logical value. Should we use viridis (\code{TRUE})
or not not (\code{FALSE})}

\item{...}{Additional arguments to customize \code{ggplot2} call}
}
\value{
A ggplot object
}
\description{
Determine evolution of Gini index
}
