% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_K.R
\name{plot_K_age}
\alias{plot_K_age}
\title{Estimate financial wealth by age}
\usage{
plot_K_age(
  simulations,
  wealth_var = "wealth",
  year_var = "annee",
  age_var = "age",
  graduation_var = "findet",
  estimation_year = 2009,
  start_year = 2009,
  final_year = 2040,
  trans = NULL,
  time_0 = c("graduation", "birth"),
  method = c("smooth", "median"),
  export_plot = FALSE,
  export_path = getwd(),
  export_filename = "06_profil_K_age",
  langage = c("French", "English"),
  xlab = "Age",
  ylab = "Simulated financial wealth (in euros)",
  size_text = 28,
  scale_viridis = FALSE,
  byvar = NULL,
  xlims = NULL
)
}
\arguments{
\item{simulations}{Longitudinal household level dataframe
storing wealth simulations}

\item{wealth_var}{Wealth variable in \code{simulations}}

\item{year_var}{Year variable}

\item{age_var}{Age variable in \code{simulations}}

\item{estimation_year}{Estimation year}

\item{start_year}{First year of simulation}

\item{final_year}{Last year of simulation}

\item{method}{Should we apply smoothing directly (\code{smooth}) or first
compute median income (\code{median}) before smoothing}

\item{export_plot}{Logical value indicating whether we want
to export the plot (\code{TRUE}, default) or not}

\item{export_path}{Path to export data}

\item{export_filename}{Filename if data are exported. PDF extension
will be added if missing}

\item{langage}{Should we label axes in English or French ?}

\item{xlab}{Label for x variable}

\item{ylab}{Label for y variable}

\item{size_text}{Parameter for text size in output}

\item{scale_viridis}{Logical value. Should we use viridis (\code{TRUE})
or not not (\code{FALSE})}

\item{byvar}{Optional variable name to separate in two figures}
}
\value{
A \code{ggplot} object. Saved on disk if \code{export_plot=TRUE}
}
\description{
Estimate financial wealth by age
}
