% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_K.R
\name{plot_rK_age}
\alias{plot_rK_age}
\alias{plot_rK_income}
\alias{plot_rK_income_1y}
\title{Estimate capital share by age or income}
\usage{
plot_rK_age(
  simulations,
  wealth_var = "rKY",
  year_var = "annee",
  age_var = "age",
  estimation_year = 2009,
  method = c("smooth", "median"),
  start_year = 2009,
  final_year = 2040,
  export_plot = FALSE,
  export_path = getwd(),
  export_filename = "04_art_rev_capital_by_age",
  langage = c("French", "English"),
  ylab = NULL,
  scale_viridis = FALSE,
  byvar = NULL,
  size_text = 32,
  xlims = NULL
)

plot_rK_income(
  simulations,
  year_var = "annee",
  income_var = "y_indiv",
  capital_income_var = "rK",
  estimation_year = 2009,
  start_year = 2009,
  final_year = 2040,
  export_plot = FALSE,
  export_path = getwd(),
  export_filename = "04_art_rev_capital_by_age",
  langage = c("French", "English"),
  scale.x = c("level", "log"),
  scale.y = c("level", "log"),
  ylab = NULL,
  size_text = 32,
  scale_viridis = FALSE,
  lower_bound_income = NULL,
  byvar = NULL
)

plot_rK_income_1y(
  simulations,
  year_var = "annee",
  income_var = "y_indiv",
  capital_income_var = "rK",
  year = 2019,
  export_plot = FALSE,
  export_path = getwd(),
  export_filename = "04_art_rev_capital_by_age",
  langage = c("French", "English"),
  scale.x = c("level", "log"),
  scale.y = c("level", "log"),
  xlab = NULL,
  ylab = NULL,
  scale_viridis = FALSE,
  lower_bound_income = NULL,
  byvar = NULL
)
}
\arguments{
\item{simulations}{Longitudinal household level dataframe
storing wealth simulations}

\item{year_var}{Year variable}

\item{estimation_year}{Estimation year}

\item{start_year}{First year of simulation}

\item{final_year}{Last year of simulation}

\item{export_plot}{Logical value indicating whether we want
to export the plot (\code{TRUE}, default) or not}

\item{export_path}{Path to export data}

\item{export_filename}{Filename if data are exported. PDF extension
will be added if missing}

\item{langage}{Should we label axes in English or French ?}

\item{ylab}{A label for y variable if default is not ok}

\item{scale_viridis}{Logical value. Should we use viridis (\code{TRUE})
or not not (\code{FALSE})}

\item{byvar}{Optional variable name to separate in two figures}

\item{size_text}{Parameter for text size in output}

\item{income_var}{Income variable to use as income definition
in \code{plot_rK_income}}

\item{capital_income_var}{Variable name storing capital income}

\item{scale.x, scale.y}{Scale for income variables. Can be \emph{level} (default)
or \emph{log} (recommanded option)}

\item{lower_bound_income}{A lower bound for income in \code{plot_rK_income}.
Ignored if \code{NULL} (default)}

\item{year}{Year to keep. Only valid for \code{plot_rK_income_1y}}

\item{xlab}{Label for x axis}
}
\value{
A \code{ggplot} object. Saved on disk if \code{export_plot=TRUE}
}
\description{
Estimate capital share by age or income
}
