% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_share_top.R
\name{plot_share_top}
\alias{plot_share_top}
\title{Plot top earners and owners shares in time}
\usage{
plot_share_top(
  simulations,
  thr = 0.9,
  year_var = "annee",
  vars = c("y_indiv", "wealth", "Y"),
  labels = c("Labor income", "Financial wealth", "Total income"),
  start_year = 2009,
  final_year = 2040,
  return_data = FALSE,
  export_plot = FALSE,
  export_path = getwd(),
  export_filename = "04_gini",
  langage = c("French", "English"),
  scale_viridis = FALSE,
  ...
)
}
\arguments{
\item{simulations}{Longitudinal household level dataframe
storing wealth simulations}

\item{thr}{Threshold for top income/wealth groups}

\item{year_var}{Year variable}

\item{vars}{Variables that should be represented}

\item{labels}{Variable labels for legend}

\item{start_year}{First year of simulation}

\item{final_year}{Last year of simulation}

\item{return_data}{Logical value indicating whether we
should return values (\code{TRUE}) or a plot
(\code{FALSE}, default)}

\item{export_plot}{Logical value indicating whether we want
to export the plot (\code{TRUE}, default) or not}

\item{export_path}{Path to export data}

\item{export_filename}{Filename if data are exported. PDF extension
will be added if missing}

\item{langage}{Should we label axes in English or French ?}

\item{scale_viridis}{Logical value. Should we use viridis (\code{TRUE})
or not not (\code{FALSE})}

\item{...}{Additional arguments to customize \code{ggplot2} call}
}
\description{
Plot the evolution of time of top owners and
 earners share
}
