% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inheritance.R
\name{predict_inheritance_old}
\alias{predict_inheritance_old}
\title{Add expected inheritance from a statistical model}
\usage{
predict_inheritance_old(household_table, data_event, tables, inheritance_model)
}
\arguments{
\item{household_table}{Longitudinal data}

\item{data_event}{Data storing inheritance events}

\item{tables}{Destinie's tables}

\item{inheritance_model}{Inheritance model}
}
\value{
\code{household_table} with two new columns:
 \code{H_received} et \code{H_given}
}
\description{
Add expected inheritance from a statistical model
}
