% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_Destinie.R
\name{prepare}
\alias{prepare}
\alias{prepare_dplyr}
\alias{prepare_dt}
\title{Prepare individual panel data before household aggregation}
\usage{
prepare(table_indiv, pensions, descript)

prepare_dplyr(table_indiv, pensions, descript)

prepare_dt(table_indiv, pensions, descript)
}
\arguments{
\item{table_indiv}{Individual data table on income (wages or pensions)}

\item{pensions}{Dataframe of pensions by id:year}

\item{descript}{Individual information dataframe}
}
\value{
Longitudinal individual data
}
\description{
Prepare individual panel data before household aggregation
}
\details{
In practice, this function handles special cases from Destinie to get a proper table
}
\examples{
\dontrun{
pre_indiv = capitulation::prepare(table_indiv = pre_indiv, pensions = pensions,
descript = description)}
}
