% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for microsimulation of capitulation accumulation}
\usage{
prepare_data(
  path_data = "./inst/dataINSEE",
  path_data_suffix = "/Destinie",
  extension = ".Rda",
  inheritance_model = NULL,
  selection_model = NULL,
  call.Rcpp = TRUE,
  estimation_year = 2009,
  id_var = "Id",
  income_var = "salaire",
  age_var = "age",
  ageliq_var = "ageliq",
  matrimonial_var = "matri",
  wealthvar_survey = "PATFISOM",
  debt_wealthSurvey = NULL,
  findet_var = "findet",
  weight_var = NULL,
  deflate = TRUE,
  quantile = 0.9,
  groupvar = "top",
  time_0 = c("graduation", "birth"),
  drawK = TRUE,
  taille_tr_age = 5,
  taille_tr_agfinetu = 2,
  ...
)
}
\arguments{
\item{path_data}{Directory where files should be found}

\item{inheritance_model}{A model that presents a `predict` method
that can be used to predict inheritance}

\item{call.Rcpp}{Logical value indicating whether we internally want to use 
\code{C++} (\code{TRUE}) or \code{R} (\code{FALSE}) language}

\item{estimation_year}{Estimation year in wealth survey}

\item{id_var}{Individual household identifier column name}

\item{income_var}{Income variable in microsimulated data}

\item{age_var}{Individual age column name}

\item{ageliq_var}{Individual liquidation age variable name}

\item{matrimonial_var}{Matrimonial status variable}

\item{wealthvar_survey}{Variable for wealth in survey data}

\item{findet_var}{End of studying year variable}

\item{weight_var}{Variable name to weight observations. 
By default, set to \code{NULL} to consider individual data}

\item{deflate}{Logical value indicating whether values should
be deflated}

\item{quantile}{Quantile used (between 0 and 1) to cut
population in two classes}

\item{groupvar}{Variable prefix for the group identification.
Variable name will be \code{group_var}_(1-\code{quantile})}

\item{drawK}{Logical value indicating whether we would like to draw observed value
(wealth, age) for people that have no wealth observed in \code{estimation_year}}

\item{...}{Additional parameters that should be passed. Currently
not used}
}
\value{
Individual table finalized, ready for microsimulation.
 It is stored as a list
}
\description{
\code{prepare_data} is a global function used to perform
 several data cleaning steps, see \code{details} section
 
 @details 
 
\code{prepare_data} executes sequentially:
\enumerate{
 \item \link{import_Destinie}, \link{arrange_Destinie}
  and \link{prepare}
   to read and arrange \code{Destinie} data into \code{R}
 \item \link{household_composition} and \link{income_household}
  to join household composition to individual information
 \item \link{match_wealthSurvey} to merge individual microsimulated
  trajectory with information observed in wealth survey
 \item Data cleaning to keep only useful information for
  life-cycle model
}
}
