% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_wealth_structural}
\alias{simulate_wealth_structural}
\alias{simulate_wealth_structural_ra}
\alias{simulate_wealth_structural_beta}
\alias{simulate_wealth_structural_beta2}
\title{Simulate Kt all along time period for a given K0 (starting wealth)}
\usage{
simulate_wealth_structural(
  K0,
  consumption0,
  income,
  UC,
  r = 0.02,
  returnLast = FALSE
)

simulate_wealth_structural_ra(
  K0,
  consumption0,
  income,
  UC,
  r = 0.02,
  gamma = 0.5,
  returnLast = FALSE
)

simulate_wealth_structural_beta(
  K0,
  consumption0,
  income,
  UC,
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  returnLast = FALSE
)

simulate_wealth_structural_beta2(
  K0,
  consumption0,
  income,
  UC,
  tau,
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  returnLast = FALSE
)
}
\arguments{
\item{K0}{Capital for the first period}

\item{consumption0}{Consumption path}

\item{income}{Income vector}

\item{UC}{Number of consumption unit in household at time t=0}

\item{r}{Exogeneous interest rate}

\item{returnLast}{Return last element}

\item{gamma}{Risk-aversion coefficient that gives instantaneous utility function curvature}

\item{beta}{Utility discount factor}

\item{tau}{Time vector since active life began}
}
\description{
Capitalize wealth given capital dynamics $K_t = (1+r)(K_{t-1}+Y_t-C_0)$,
 initial capital (K0), consumption (C0) and income dynamics. Different
 functions that are associated with different models
 \itemize{
    \item \code{simulate_wealth_structural}: simplest life cycle model where
      \eqn{\beta =1} and no risk aversion
    \item \code{simulate_wealth_structural}: life cycle model with
      \eqn{\beta =1 } and risk aversion
    \item \code{simulate_wealth_structural}: complete life cycle model with
      \eqn{\beta \leq 1} and risk aversion
 }
}
\seealso{
\link{calibrate_K0}
}
