% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_wealth_structural.R
\name{starting_wealth_structural}
\alias{starting_wealth_structural}
\title{[Discarded] Determine first generation wealth given observation in
 survey data (structural approach)
 
 This function is discarded}
\usage{
starting_wealth_structural(
  data,
  deflate = TRUE,
  wealthvar_survey = "PATFISOM",
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  weight_var = NULL
)
}
\arguments{
\item{data}{Household level longitudinal trajectories}

\item{deflate}{Logical value indicating whether we want
to use nominal (\code{FALSE}) or real (\code{TRUE}, default) series}

\item{wealthvar_survey}{Wealth variable in household survey}

\item{r}{Exogeneous interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Discount factor}

\item{weight_var}{Variable storing individual weight inside household
(necessary to ensure consistency between household and
individual levels)}
}
\description{
[Discarded] Determine first generation wealth given observation in
 survey data (structural approach)
 
 This function is discarded
}
