% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_range}
\alias{subset_range}
\title{Wrapper for vector subsetting in Rcpp}
\usage{
subset_range(x, start = 0L, end = 1L)
}
\arguments{
\item{x}{Numeric vector}

\item{start}{Index start. Indexation follows C++ rule hence begins at 0}

\item{end}{Index of ending. Indexation follows C++ rule hence ends at n-1}
}
\value{
x values between \code{start} and \code{end}
}
\description{
Wrapper for vector subsetting in Rcpp
}
