% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_shares.R
\name{table_shares}
\alias{table_shares}
\title{Automatize table for top income/wealth shares}
\usage{
table_shares(
  simulations,
  thr = c(0.9, 0.99),
  vars = c("y_indiv", "wealth", "Y"),
  labels = c("Labor income", "Financial wealth", "Total income"),
  year = 2015L,
  year_var = "annee",
  filename = NULL
)
}
\arguments{
\item{simulations}{Longitudinal household level dataframe
storing wealth simulations}

\item{thr}{Threshold for top income/wealth groups}

\item{vars}{Variables that should be represented}

\item{labels}{Variable labels for legend}

\item{year}{Year that should be used}

\item{year_var}{Year variable}

\item{filename}{A filename (including directory path).
If \code{NULL}, table is returned and file is not
written}
}
\value{
Depending on the value of \code{filename}
}
\description{
Automatize table for top income/wealth shares
}
