% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_household.R
\name{transform_household}
\alias{transform_household}
\title{Transform individual data into household data}
\usage{
transform_household(
  table_indiv,
  id_var = "Id",
  income_var = "salaire",
  age_var = "age",
  ageliq_var = "ageliq",
  matrimonial_var = "matri"
)
}
\arguments{
\item{table_indiv}{Individual level data where
people are matched to husband/spouse id}

\item{id_var}{Individual household identifier column name}

\item{income_var}{Individual income column name}

\item{age_var}{Individual age column name}

\item{ageliq_var}{Individual liquidation age variable name}

\item{matrimonial_var}{Matrimonial status variable}
}
\value{
Individual level data with income aggregated
 at household level
}
\description{
Finalize transformation of individual data into household data.
Aggregate income at household level data and keep only
 one adult by household. It is assumed \link{assign_referent}
 has already been called.
}
\details{
By assumption, male is assumed to be the household head.

Several arguments are proposed to ensure a column name discrepancy
 between two inputs will not produce different results
}
\examples{
# Toy example
z <- data.frame(
  Id = rep(1:2,2),
  annee = c(2008,2008,2009,2009),
  sexe = rep(1:2,2),
  # Married together
  matri = rep(2,4),
  conjoint = rep(2:1,2),
  # Male is household head
  referent = rep(1,4),
  ageliq = rep(c(60,65),2),
  salaire = c(10000,2,0,8),
  age = c(25,23,26,24),
  findet = rep(20,4)
)

# Results are consistent in dplyr or dt
# approaches
z_dplyr <- z
z_dt <- data.table::as.data.table(z)
head(z_dplyr)
head(z_dt)
}
