% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnestDT.R
\name{unnest}
\alias{unnest}
\alias{unnest.data.table}
\title{A small method to unnest objects}
\usage{
unnest(data, ...)

\method{unnest}{data.table}(data, ...)
}
\arguments{
\item{data}{A `data.table` object}

\item{...}{Additional arguments}
}
\value{
A character vector with names
 that can be passed to latex
}
\description{
A small method to unnest objects
}
