% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wealth_accumulation.R
\name{wealth_accumulation}
\alias{wealth_accumulation}
\title{Simulate wealth dynamics at household level}
\usage{
wealth_accumulation(
  household_table,
  years = 2009:2070,
  start_year = 2009,
  r = 0.02,
  gamma = 0.5,
  beta = 1,
  verbose = TRUE,
  age_var = "age",
  findet_var = "findet",
  income_var = "salaire_tot",
  wealthvar_survey = "PATFISOM",
  weight_var = NULL,
  return_split = FALSE,
  return_last = FALSE
)
}
\arguments{
\item{household_table}{Longitudinal household table}

\item{years}{Sequence of years to consider. By default from 2009 to 2100}

\item{start_year}{Year where microsimulation starts}

\item{r}{Exogeneous interest rate}

\item{gamma}{Risk aversion coefficient}

\item{beta}{Discount factor}

\item{verbose}{Logical value indicating whether we want to print some messages
during computations}

\item{age_var}{Individual age column name}

\item{findet_var}{Variable name storing end of year studying information}

\item{income_var}{Individual income column name}

\item{wealthvar_survey}{Variable for wealth in survey data}

\item{weight_var}{Variable storing individual weight inside household. This can be 1
(individual is the unique adult representing household), 2 (couples) or
a value related to consumption units}

\item{return_split}{Logical value indicating whether we want to return pivot
years where family change ocurred}

\item{return_last}{Logical value indicating whether we want to return
\eqn{Kt} sequence up to `T-1` (`FALSE`, default) or `T` (`TRUE`)}
}
\value{
Dataframe with wealth variable simulated
}
\description{
Main function of \code{capitulation} package to simulate
 wealth accumulation across time. Handle family breaks and marriage
 as well as capitalize wealth given behavioral assumptions.
}
